refactor(core): extract FFI parsing helpers
